requirements.txt
amqp==2.5.2
celery==4.3.0
click==8.0.1
flask-compress==1.4.0
flask-restful==0.3.9
flask==2.0.1
werkzeug==2.0.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.1.0
inject==4.3.1
pymysql==0.9.3
redis==3.2.1
requests==2.26.0
sqlalchemy==1.4.27
urllib3==1.26.6
xlwt==1.3.0
dacite==1.6.0
# mysqlclient==2.0.1
numpy==1.19.5
marshmallow==3.13.0
cryptography==3.4.7
sentry_sdk==1.4.3
celery-redbeat==2.0.0
sqlalchemy_utils==0.37.8
python-redis-lock==3.7.0
